+++ src/generateReport.ts
@@ -1,5 +1,5 @@
 import { AuditResults } from './audit';
-import { Advisories, Statistics } from './types';
+import { Advisories, SeverityCountsWithTotal, Statistics } from './types';
 
 export interface AuditReport {
   statistics: Statistics;
@@ -11,6 +11,54 @@ export interface AuditReport {
   // "advisory|path" that are marked as ignored, but not in the results
   missing: readonly string[];
 }
+
+const severityCountsWithTotal: SeverityCountsWithTotal = {
+  total: 0,
+  info: 0,
+  low: 0,
+  moderate: 0,
+  high: 0,
+  critical: 0
+};
+
+const generateStatistics = (
+  ignores: readonly string[],
+  results: AuditResults
+): Statistics => {
+  const statistics: Statistics = {
+    dependencies: results.dependencyStatistics,
+    severities: { ...severityCountsWithTotal },
+    vulnerable: { ...severityCountsWithTotal },
+    ignored: { ...severityCountsWithTotal }
+  };
+
+  Object.values(results.advisories).forEach(({ id, findings, severity }) => {
+    const { vulnerable, ignored } = findings.reduce(
+      (sums, { paths }) => {
+        const { length: count } = paths.filter(path =>
+          ignores.includes(`${id}|${path}`)
+        );
+
+        sums.vulnerable += paths.length - count;
+        sums.ignored += count;
+
+        return sums;
+      },
+      { vulnerable: 0, ignored: 0 }
+    );
+
+    statistics.severities[severity] += ignored + vulnerable;
+    statistics.severities.total += ignored + vulnerable;
+
+    statistics.vulnerable[severity] += vulnerable;
+    statistics.vulnerable.total += vulnerable;
+
+    statistics.ignored[severity] += ignored;
+    statistics.ignored.total += ignored;
+  });
+
+  return statistics;
+};
 
 export const generateReport = (
   ignores: readonly string[],
@@ -44,5 +92,11 @@ export const generateReport = (
       [[], [], [...ignores]]
     );
 
-  return { ...results, vulnerable, ignored, missing };
+  return {
+    advisories: results.advisories,
+    statistics: generateStatistics(ignores, results),
+    vulnerable,
+    ignored,
+    missing
+  };
 };